// PortalAutoservicio/src/app/models/datos-personales.ts
export class DatosPersonales {
    CLAVESERVIDOR: string;
    APELLIDOPATERNO: string;
    APELLIDOMATERNO: string;
    MOMBRE: string;
    NOMBRECOMPLETO: string;
    IDGENERO: string;
    NOMBREGENERO: string;
    IDPAIS: string;
    NOMBREPAIS: string;
    IDTRATAMIENTO: string;
    NOMBRETRATAMIENTO: string;
    FECHANACIMIENTO: string;
    RFC: string;
    CURP: string;
    IDESCOLARIDAD: string;
    DESCRIESCOLARIDAD: string;
    IDESTADOCIVIL: string;
    NOMBREESTADOCIVIL: string;
    COMENTARIOS: string;
    DESEMPENO: string;
    KPI: string;
    ESCALAFON: string;
    CONSULTA: string;
    IDROL: number;
    BOACTIVO: string;
    BOPRORROGA: string;
    IDPUESTO: string;
    NPUESTO: string;

}